# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for calculating order pack breakdowns.",
        "title": "Orders Calculation API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Lucas Lemos",
            "email": "lucashenriqueblemos@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://mit-license.org/"
        },
        "version": "1.0"
    },
    "host": "orders-calculation.luk3skyw4lker.com",
    "basePath": "/",
    "paths": {
        "/orders": {
            "get": {
                "description": "Retrieve a list of all orders",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Order"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/payload.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/payload.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create an order with the specified number of items",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create an order",
                "parameters": [
                    {
                        "description": "the order to be created",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.CreateOrder"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/payload.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/payload.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/{order_id}": {
            "get": {
                "description": "Retrieve the details of an order using its ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get an order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ID of the order to retrieve",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/payload.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/payload.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/payload.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pack-sizes": {
            "get": {
                "description": "Retrieve a list of all available pack sizes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PackSizes"
                ],
                "summary": "Get all pack sizes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PackSize"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/payload.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new pack size to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PackSizes"
                ],
                "summary": "Create a new pack size",
                "parameters": [
                    {
                        "description": "The pack size to create",
                        "name": "packSize",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.CreatePackSize"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.PackSize"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/payload.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/payload.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pack-sizes/{pack_size_id}": {
            "put": {
                "description": "Update the size of an existing pack size",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PackSizes"
                ],
                "summary": "Update an existing pack size",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ID of the pack size to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The updated pack size data",
                        "name": "packSize",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.UpdatePackSize"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PackSize"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/payload.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/payload.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Order": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "items_count": {
                    "type": "integer"
                },
                "pack_setup": {
                    "type": "string"
                }
            }
        },
        "models.PackSize": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "payload.CreateOrder": {
            "type": "object",
            "required": [
                "items_count"
            ],
            "properties": {
                "items_count": {
                    "type": "integer"
                }
            }
        },
        "payload.CreatePackSize": {
            "type": "object",
            "required": [
                "size"
            ],
            "properties": {
                "size": {
                    "type": "integer"
                }
            }
        },
        "payload.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "payload.UpdatePackSize": {
            "type": "object",
            "required": [
                "id",
                "size"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        }
    }
}